knowledge-hub/apps/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.27.2
celery==5.3.4
rq==1.16.2
pymupdf==1.24.9
docx2txt==0.8
faiss-cpu==1.8.0
sentence-transformers==2.7.0
numpy==1.24.3
